+++ src/library/DIPS.Mobile.UI/Components/Pickers/DateAndTimePicker/iOS/DateAndTimePickerHandler.cs
@@ -1,63 +1,29 @@
-using CoreGraphics;
-using DIPS.Mobile.UI.API.Library;
-using DIPS.Mobile.UI.Components.Pickers.DatePicker.iOS;
-using DIPS.Mobile.UI.Components.Pickers.Platforms.iOS;
+using DIPS.Mobile.UI.Components.Pickers.DatePickerShared.iOS;
 using DIPS.Mobile.UI.Platforms.iOS;
 using Foundation;
-using Microsoft.Maui.Handlers;
 using UIKit;
 
 // ReSharper disable once CheckNamespace
 namespace DIPS.Mobile.UI.Components.Pickers.DateAndTimePicker;
 
-public partial class DateAndTimePickerHandler : ViewHandler<DateAndTimePicker, DUIDatePicker>
+public partial class DateAndTimePickerHandler : BaseDatePickerHandler
 {
-    private bool m_isOpen;
-
     protected override DUIDatePicker CreatePlatformView()
     {
         return new DUIDatePicker
         {
-            Mode = UIDatePickerMode.DateAndTime, PreferredDatePickerStyle = UIDatePickerStyle.Compact
+            Mode = UIDatePickerMode.DateAndTime,
+            PreferredDatePickerStyle = UIDatePickerStyle.Compact
         };
     }
-    
-    protected override void ConnectHandler(DUIDatePicker platformView)
-    {
-        base.ConnectHandler(platformView);
 
-        platformView.ValueChanged += OnDateSelected;
-        platformView.SetInLineLabelColors();
-        
-        platformView.EditingDidBegin += OnOpen;
-        platformView.EditingDidEnd += OnClose;
-        DUI.OnRemoveViewsLocatedOnTopOfPage += TryClose;
-    }
-
-    private void OnOpen(object? sender, EventArgs e)
+    protected override void OnValueChanged(object? sender, EventArgs e)
     {
-        m_isOpen = true;
-    }
-    
-    private void OnClose(object? sender, EventArgs e)
-    {
-        m_isOpen = false;
-    }
-
-    private void TryClose()
-    {
-        if (!m_isOpen)
+        if(VirtualView is not DateAndTimePicker dateAndTimePicker)
             return;
         
-        var currentPresentedUiViewController = Platform.GetCurrentUIViewController();
-        currentPresentedUiViewController?.DismissViewController(false, null);
-    }
-
-    private partial void AppendPropertyMapper()
-    {
-        DateAndTimePickerPropertyMapper.Add(nameof(DateAndTimePicker.HorizontalOptions), MapHorizontalOptions);
-        DateAndTimePickerPropertyMapper.Add(nameof(DateAndTimePicker.Background), MapOverrideBackground);
-        DateAndTimePickerPropertyMapper.Add(nameof(DateAndTimePicker.IgnoreLocalTime), MapIgnoreLocalTime);
+        dateAndTimePicker.SelectedDateTime = (DateTime)PlatformView.Date;
+        dateAndTimePicker.SelectedDateTimeCommand?.Execute(null);
     }
 
     private static partial void MapMaximumDate(DateAndTimePickerHandler handler, DateAndTimePicker dateAndTimePicker)
@@ -76,21 +42,6 @@ public partial class DateAndTimePickerHandler : ViewHandler<DateAndTimePicker, D
         handler.PlatformView.MinimumDate = ((DateTime)dateAndTimePicker.MinimumDate).ConvertDate();
     }
 
-    private static void MapOverrideBackground(DateAndTimePickerHandler handler, DateAndTimePicker dateAndTimePicker)
-    {
-    }
-
-    private static void MapHorizontalOptions(DateAndTimePickerHandler handler, DateAndTimePicker dateAndTimePicker)
-    {
-        handler.PlatformView.SetHorizontalAlignment(dateAndTimePicker);
-    }
-
-    private void OnDateSelected(object? sender, EventArgs e)
-    {
-        VirtualView.SelectedDateTime = (DateTime)PlatformView.Date;
-        VirtualView.SelectedDateTimeCommand?.Execute(null);
-    }
-
     private static partial void MapIgnoreLocalTime(DateAndTimePickerHandler handler, DateAndTimePicker dateAndTimePicker)
     {
         handler.PlatformView.TimeZone = dateAndTimePicker.IgnoreLocalTime ? new NSTimeZone("UTC") : NSTimeZone.LocalTimeZone;
@@ -99,16 +50,5 @@ public partial class DateAndTimePickerHandler : ViewHandler<DateAndTimePicker, D
     private static partial void MapSelectedDate(DateAndTimePickerHandler handler, DateAndTimePicker dateAndTimePicker)
     {
         handler.PlatformView.SetDate(dateAndTimePicker.SelectedDateTime.ConvertDate(), true);
-    }
-
-    protected override void DisconnectHandler(DUIDatePicker platformView)
-    {
-        base.DisconnectHandler(platformView);
-
-        platformView.DisposeLayer();
-        platformView.ValueChanged -= OnDateSelected;
-        platformView.EditingDidBegin -= OnOpen;
-        platformView.EditingDidEnd -= OnClose;
-        DUI.OnRemoveViewsLocatedOnTopOfPage -= TryClose;
     }
 }