+++ src/library/DIPS.Mobile.UI/Components/Pickers/DatePicker/Inline/iOS/InlineDatePicker.cs
@@ -1,7 +1,7 @@
 namespace DIPS.Mobile.UI.Components.Pickers.DatePicker.Inline.iOS;
 
 /// <summary>
-/// Internal because we only support it on iOS until we find a Android version of it
+/// Internal because we only support it on iOS until we find an Android version of it
 /// </summary>
 internal class InlineDatePicker : DatePicker
 {