+++ src/library/DIPS.Mobile.UI/Components/Pickers/DatePickerShared/IDatePicker.cs
@@ -1,0 +1,6 @@
+namespace DIPS.Mobile.UI.Components.Pickers.DatePickerShared;
+
+public interface IDatePicker : IView
+{
+   
+}