+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-## [28.5.1]
+## [28.6.1] 
 - [iOS][Chip] Add slightly more width when `IsToggleable=True` in order to not truncate title.
+
+## [28.6.0] 
+- Added new components: `NullableDatePicker`, `NullableTimePicker` and `NullableDateAndTimePicker`.
+- [iOS] DatePickers will now have exact same design as Chips. 
 
 ## [28.5.0] 
 - Resources was updated from DIPS.Mobile.DesignTokens